Remove feature to view question display
Non-admin users that are not matched were able to enter the question page previously without finding a match. However, now that we intend that the question page is only displayed upon finding a suitable match, we removed this feature. Users can still see and read the entire question by reading from the display. Hence we find this feature of redirecting them to a dedicated page to be redundant as well. This is committed in this pr as this would lead to bugs and erroneous behavior for collaboration service